+++ src/Instance/HoldingsList/consortium/MemberTenantHoldings/MemberTenantHoldings.test.js
@@ -1,6 +1,7 @@
 import { BrowserRouter as Router } from 'react-router-dom';
 
 import { screen } from '@folio/jest-config-stripes/testing-library/react';
+import { useUserTenantPermissions } from '@folio/stripes/core';
 
 import '../../../../../test/jest/__mock__';
 import {
@@ -9,10 +10,22 @@ import {
 } from '../../../../../test/jest/helpers';
 
 import { instance } from '../../../../../test/fixtures';
+import { useHoldingsFromStorage } from '../../../../hooks';
 
 import MemberTenantHoldings from './MemberTenantHoldings';
 
-jest.mock('../../../../hooks/useMemberTenantHoldings', () => () => ({ holdings: [{}], isLoading: false }));
+jest.mock('../../../../hooks/useMemberTenantHoldings', () => () => ({ holdings: [{ id: 'holdingId' }], isLoading: false }));
+jest.mock('@folio/stripes/core', () => ({
+  ...jest.requireActual('@folio/stripes/core'),
+  useUserTenantPermissions: jest.fn().mockReturnValue({
+    userPermissions: [],
+    isFetching: false,
+  }),
+}));
+jest.mock('../../../../hooks', () => ({
+  ...jest.requireActual('../../../../hooks'),
+  useHoldingsFromStorage: jest.fn(),
+}));
 jest.mock('../../HoldingsList', () => () => <>Holdings</>);
 jest.mock('../LimitedHoldingsList', () => ({
   ...jest.requireActual('../LimitedHoldingsList'),
@@ -28,32 +41,25 @@ const mockMemberTenant = {
   name: 'College',
 };
 
+const accordionId = `${mockMemberTenant.id}.${instance.id}`;
+
 const userTenantFullPermissions = [{
-  tenantId: 'college',
-  permissionNames: [{
-    permissionName: 'ui-inventory.holdings.create',
-    subPermissions: ['test subPermission 1']
-  }, {
-    permissionName: 'ui-inventory.instance.view',
-    subPermissions: ['test subPermission 1']
-  }, {
-    permissionName: 'ui-inventory.item.create',
-    subPermissions: ['test subPermission 1']
-  }],
+  permissionName: 'ui-inventory.holdings.create',
+  subPermissions: ['test subPermission 1']
+}, {
+  permissionName: 'ui-inventory.instance.view',
+  subPermissions: ['test subPermission 1']
+}, {
+  permissionName: 'ui-inventory.item.create',
+  subPermissions: ['test subPermission 1']
 }];
 
-const userTenantLimitedPermissions = [{
-  tenantId: 'college',
-  permissionNames: [],
-}];
-
-const renderMemberTenantHoldings = ({ userTenantPermissions }) => {
+const renderMemberTenantHoldings = () => {
   const component = (
     <Router>
       <MemberTenantHoldings
         instance={instance}
         memberTenant={mockMemberTenant}
-        userTenantPermissions={userTenantPermissions}
       />
     </Router>
   );
@@ -62,15 +68,30 @@ const renderMemberTenantHoldings = ({ userTenantPermissions }) => {
 };
 
 describe('MemberTenantHoldings', () => {
+  beforeEach(() => {
+    useHoldingsFromStorage.mockClear().mockReturnValue(
+      [{ [accordionId]: true }, jest.fn()]
+    );
+  });
+
+  afterAll(() => {
+    jest.clearAllMocks();
+  });
+
   it('should render member tenant accordion', () => {
-    renderMemberTenantHoldings({ userTenantPermissions: userTenantFullPermissions });
+    renderMemberTenantHoldings();
 
     expect(screen.getByText('College')).toBeInTheDocument();
   });
 
   describe('tenant with inventory permissions', () => {
     it('should render member tenant\'s holdings', () => {
-      renderMemberTenantHoldings({ userTenantPermissions: userTenantFullPermissions });
+      useUserTenantPermissions.mockClear().mockReturnValue({
+        userPermissions: userTenantFullPermissions,
+        isFetching: false,
+      });
+
+      renderMemberTenantHoldings();
 
       expect(screen.getByText('Holdings')).toBeInTheDocument();
     });
@@ -78,14 +99,24 @@ describe('MemberTenantHoldings', () => {
 
   describe('tenant with limited permissions', () => {
     it('should render member tenant\'s holdings with limited information', () => {
-      renderMemberTenantHoldings({ userTenantPermissions: userTenantLimitedPermissions });
+      useUserTenantPermissions.mockClear().mockReturnValue({
+        userPermissions: [],
+        isFetching: false,
+      });
+
+      renderMemberTenantHoldings();
 
       expect(screen.getByText('LimitedHoldingsList')).toBeInTheDocument();
     });
   });
 
   it('should render Add holdings button', () => {
-    renderMemberTenantHoldings({ userTenantPermissions: userTenantFullPermissions });
+    useUserTenantPermissions.mockClear().mockReturnValue({
+      userPermissions: userTenantFullPermissions,
+      isFetching: false,
+    });
+
+    renderMemberTenantHoldings();
 
     expect(screen.getByRole('button', { name: 'Add holdings' })).toBeInTheDocument();
   });